feat: Learn about context API in React
Learn about why do we need Context API in React, what problems it solves, how to use it, and then convert our Attendance app with Context API
