+++ src/index.js
@@ -2,13 +2,18 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
+import StudentProvider from "./contexts/Student";
 
 import reportWebVitals from "./reportWebVitals";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
 	<React.StrictMode>
-		<App />
+		<StudentProvider>
+			<App>
+				<h2>Hello</h2>
+			</App>
+		</StudentProvider>
 	</React.StrictMode>,
 );
